Renamed to client to better represent function
